step#2 - 노말/라스트프레임 중복제거 시도
